requirements.txt
flask
docker
pyyaml